requirements.txt
asgiref==3.8.1
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
future==1.0.0
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.1.0
rapidfuzz==3.9.2
requests==2.32.3
requests-toolbelt==1.0.0
serializers==0.2.4
shellingham==1.5.4
sqlparse==0.5.0
tomlkit==0.12.5
trove-classifiers==2024.5.22
urllib3==2.2.1
virtualenv==20.26.2
xattr==1.1.0
